requirements.txt
azure-iot-device
azure-storage-blob
streamlit
time